refactor: move find_git_root to var_expansion
Also remove var_expansion_helpers.

- Consolidated the find_git_root function into utils.var_expansion,
  removing the now redundant var_expansion_helpers module.
- Updated imports in compile_handling and test_var_expansion to reflect
  this change.
