+++ src/utils/__init__.py
@@ -3,8 +3,8 @@
 from .var_expansion import (
     expand_variables,
     expand_variables_in_list,
+    find_git_root,
 )
-from .var_expansion_helpers import find_git_root
 
 __all__ = [
     "expand_variables",
